feat(DataMeter): use translate instead of width to show progress
