+++ components/core/DataMeter.js
@@ -1,109 +1,35 @@
 import * as React from "react";
-import * as Constants from "~/common/constants";
-import * as Strings from "~/common/strings";
 
 import { css } from "@emotion/react";
+import { motion } from "framer-motion";
 
-const STYLES_CONTAINER = css`
+const STYLES_DATAMETER_WRAPPER = (theme) => css`
+  width: 100%;
+  height: 8px;
+  background-color: ${theme.semantic.bgBlurWhiteTRN};
+  border: 1px solid ${theme.semantic.borderGrayLight4};
   border-radius: 4px;
-  box-shadow: 0 0 0 1px ${Constants.semantic.borderGrayLight} inset, ${Constants.shadow.lightSmall};
-  padding: 32px;
-  max-width: 100%;
-  width: 100%;
-
-  @media (max-width: ${Constants.sizes.mobile}px) {
-    padding: 24px;
-  }
-`;
-
-const STYLES_DATA = css`
-  width: 100%;
-  display: flex;
-  align-items: center;
-  height: 16px;
-  border-radius: 3px;
-  background-color: ${Constants.semantic.bgLight};
   overflow: hidden;
 `;
 
-const STYLES_DATA_METER = css`
-  flex-shrink: 0;
-  height: 16px;
-  background-color: ${Constants.system.blue};
+const STYLES_DATAMETER = (theme) => css`
+  position: relative;
+  left: -100%;
+  height: 100%;
+  background-color: ${theme.system.blue};
+  border-radius: 4px;
 `;
 
-const STYLES_ROW = css`
-  display: flex;
-  align-items: flex-end;
-  justify-content: space-between;
-  font-family: ${Constants.font.code};
-  color: ${Constants.system.grayLight2};
-  font-size: 10px;
-  margin-top: 2px;
-  text-transform: uppercase;
-`;
-
-const STYLES_LEFT = css`
-  min-width: 10%;
-  width: 100% "";
-`;
-
-const STYLES_RIGHT = css`
-  flex-shrink: 0;
-`;
-
-const STYLES_TITLE = css`
-  font-family: ${Constants.font.medium};
-  font-size: ${Constants.typescale.lvl1};
-  display: block;
-  margin-bottom: 4px;
-  overflow-wrap: break-word;
-`;
-
-const STYLES_NOTE = css`
-  margin-top: 8px;
-  font-family: ${Constants.font.text};
-  font-size: ${Constants.typescale.lvl0};
-  color: ${Constants.system.grayLight2};
-  display: block;
-  margin-bottom: 4px;
-`;
-
-export const DataMeterBar = (props) => {
-  const percentage = props.bytes / props.maximumBytes;
-
+export default function DataMeter({ bytes = 1000, maximumBytes = 4000, css, ...props }) {
+  const percentage = bytes / maximumBytes;
   return (
-    <React.Fragment>
-      <div css={STYLES_ROW}>
-        <div css={STYLES_LEFT} style={{ color: props.failed ? Constants.system.red : null }}>
-          {props.leftLabel}
-        </div>
-        <div css={STYLES_RIGHT}>{props.rightLabel}</div>
-      </div>
-
-      <div
-        css={STYLES_DATA}
-        style={{
-          marginTop: 8,
-          backgroundColor: props.failed ? Constants.system.red : null,
-        }}
-      >
-        <div css={STYLES_DATA_METER} style={{ width: `${percentage * 100}%` }} />
-      </div>
-    </React.Fragment>
-  );
-};
-
-export const DataMeter = (props) => {
-  return (
-    <div css={STYLES_CONTAINER} style={props.style}>
-      <div css={STYLES_TITLE}>
-        {Strings.bytesToSize(props.stats.bytes)} of {Strings.bytesToSize(props.stats.maximumBytes)}{" "}
-        used
-      </div>
-      <DataMeterBar bytes={props.stats.bytes} maximumBytes={props.stats.maximumBytes} />
+    <div css={[STYLES_DATAMETER_WRAPPER, css]} {...props}>
+      <motion.div
+        initial={false}
+        css={STYLES_DATAMETER}
+        animate={{ x: `${percentage * 100}%` }}
+        transition={{ type: "spring", stiffness: 170, damping: 26 }}
+      />
     </div>
   );
-};
-
-export default DataMeter;
+}
